Crate tidy

Library used by tidy and other tools.

This library contains the tidy lints and exposes it to be used by tools.

Modules

alphabetical

Checks that a list of items is in alphabetical order

bins

Tidy check to ensure that there are no binaries checked into the source tree by accident.

debug_artifacts

Tidy check to prevent creation of unnecessary debug artifacts while running tests.

deps

Checks the licenses of third-party dependencies.

edition

Tidy check to ensure that crate edition is ‘2018’ or ‘2021’.

error_codes

Tidy check to ensure error codes are properly documented and tested.

ext_tool_checks

Optional checks for file types other than Rust source

extdeps

Check for external package sources. Allow only vendorable packages.

features

Tidy check to ensure that unstable features are all in order.

fluent_alphabetical

Checks that all Flunt files have messages in alphabetical order

mir_opt_tests

Tidy check to ensure that mir opt directories do not have stale files or dashes in file names

pal

Tidy check to enforce rules about platform-specific code in std.

primitive_docs

Tidy check to make sure library/{std,core}/src/primitive_docs.rs are the same file. These are different files so that relative links work properly without having to have CARGO_PKG_NAME set, but conceptually they should always be the same.

rustdoc_gui_tests

Tidy check to ensure that rustdoc GUI tests start with a small description.

style

Tidy check to enforce various stylistic guidelines on the Rust codebase.

target_specific_tests

Tidy check to ensure that all target specific tests (those that require a --target flag) also require the pre-requisite LLVM components to run.

tests_placement

ui_tests

Tidy check to ensure below in UI test directories:

unit_tests

Tidy check to ensure #[test] and #[bench] are not used directly inside core.

unstable_book

walk

x_version

Macros

t

A helper macro to unwrap a result except also print out details like:

tidy_error 🔒

Functions

tidy_error 🔒