Module tidy::bins

Tidy check to ensure that there are no binaries checked into the source tree by accident.

In the past we’ve accidentally checked in test binaries and such which add a huge amount of bloat to the Git history, so it’s good to just ensure we don’t do that again.

Modules

os_impl 🔒

Functions

check

check_filesystem_support