tidy::extra_checks

Function get_or_create_venv

fn get_or_create_venv(
    venv_path: &Path,
    src_reqs_path: &Path,
) -> Result<PathBuf, Error>

Create a virtuaenv at a given path if it doesn’t already exist, or validate the install if it does. Returns the path to that venv’s python executable.