tidy::extra_checks

Struct ExtraCheckArg

struct ExtraCheckArg {
    auto: bool,
    lang: ExtraCheckLang,
    kind: Option<ExtraCheckKind>,
}

Fields
auto: boollang: ExtraCheckLangkind: Option<ExtraCheckKind>

None = run all extra checks for the given lang

Implementations

impl ExtraCheckArg

fn matches(&self, lang: ExtraCheckLang, kind: ExtraCheckKind) -> bool

fn is_non_auto_or_matches(&self, filepath: &str) -> bool

Returns false if this is an auto arg and the passed filename does not trigger the auto rule

fn has_supported_kind(&self) -> bool