tidy
Module features

Tidy check to ensure that unstable features are all in order.

This check will ensure properties like:

All stability attributes look reasonably well formed.

The set of library features is disjoint from the set of language features.

Library features have at most one stability level.

Library features have at most one since value.

All unstable lang features have tests to ensure they are actually unstable.

Language features in a group are sorted by feature name.

Modules

version ğŸ”’

Structs

CollectedFeatures

Feature

Enums

Status

Constants

FEATURE_GROUP_END_PREFIX ğŸ”’

FEATURE_GROUP_START_PREFIX ğŸ”’

Functions

check

collect_lang_features

collect_lang_features_in ğŸ”’

collect_lib_features

find_attr_val ğŸ”’

format_features ğŸ”’

get_and_check_lib_features ğŸ”’

get_version_and_channel ğŸ”’

map_lib_features ğŸ”’

test_filen_gate ğŸ”’

Type Aliases

Features