tidy
Module alphabetical

Checks that a list of items is in alphabetical order

Use the following marker in the code:

// tidy-alphabetical-start
fn aaa() {}
fn eee() {}
fn z() {}
// tidy-alphabetical-end

The following lines are ignored:

Empty lines

Lines that are indented with more or less spaces than the first line

Lines starting with //, # (except those starting with #!), ), ], } if the comment has the same indentation as the first line

Lines starting with a closing delimiter (), [, }) are ignored.

If a line ends with an opening delimiter, we effectively join the following line to it before checking it. E.g. foo(\nbar) is treated like foo(bar).

Constants

END_MARKER 🔒

START_MARKER 🔒

Functions

check

check_lines 🔒

check_section 🔒

indentation 🔒

is_close_bracket 🔒