Module tidy::deps

Checks the licenses of third-party dependencies.

Constants

EXCEPTIONS 🔒

These are exceptions to Rust’s permissive licensing policy, and should be considered bugs. Exceptions are only allowed in Rust tooling. It is crucial that no exception crates be dependencies of the Rust runtime (std/test).

EXCEPTIONS_BOOTSTRAP 🔒

EXCEPTIONS_CARGO 🔒

EXCEPTIONS_CRANELIFT 🔒

LICENSES 🔒

These are licenses that are allowed for all crates, including the runtime, rustc, tools, etc.

PERMITTED_CRANELIFT_DEPENDENCIES 🔒

PERMITTED_DEPS_LOCATION 🔒

PERMITTED_RUSTC_DEPENDENCIES 🔒

Crates rustc is allowed to depend on. Avoid adding to the list if possible.

RUNTIME_CRATES 🔒

These are the root crates that are part of the runtime. The licenses for these and all their dependencies must not be in the exception list.

Functions

check

Dependency checks.

check_license_exceptions 🔒

Check that all licenses are in the valid list in LICENSES.

check_permitted_dependencies 🔒

Checks the dependency of restricted_dependency_crates at the given path. Changes bad to true if a check failed.

check_rustfix 🔒

compute_runtime_crates 🔒

Finds all the packages that are in the rust runtime.

deps_of_filtered 🔒

Recursively find all dependencies.

direct_deps_of 🔒

pkg_from_id 🔒

pkg_from_name 🔒

Finds a package with the given name.