tidy
Module unstable_book

Constants

COMPILER_FLAGS_DIR

LANG_FEATURES_DIR

LIB_FEATURES_DIR

PATH_STR

Functions

check

collect_unstable_book_lang_features_section_file_names 🔒

Retrieves file names of all library feature sections in the Unstable Book with:

collect_unstable_book_lib_features_section_file_names 🔒

Retrieves file names of all language feature sections in the Unstable Book with:

collect_unstable_book_section_file_names

collect_unstable_feature_names

Retrieves names of all unstable features.

dir_entry_is_file 🔒

Tests whether DirEntry is a file.

unstable_book_lang_features_path

Builds the path to the directory where the features are documented within the Unstable Book source directory.

unstable_book_lib_features_path

Builds the path to the directory where the features are documented within the Unstable Book source directory.

unstable_book_path

Builds the path to the Unstable Book source directory from the Rust ‘src’ directory.